+++ support/python/src/nebula_sdk/outputs.py
@@ -1,15 +1,18 @@
 import json
+from typing import Any
 from urllib.parse import quote
+
+from requests import Session
 
 from .util import JSONEncoder
 
 
-class Outputs(object):
+class Outputs:
 
-    def __init__(self, client):
+    def __init__(self, client: Session) -> None:
         self._client = client
 
-    def set(self, name, value):
+    def set(self, name: str, value: Any) -> None:
         r = self._client.put(
             'http+api://api/outputs/{0}'.format(quote(name)),
             data=json.dumps(value, cls=JSONEncoder),